DS-146: minor refactoring
